+++ gateway-service/src/main/java/com/ca/mfaas/gateway/filters/pre/LocationFilter.java
@@ -16,7 +16,6 @@ import com.netflix.zuul.ZuulFilter;
 import com.netflix.zuul.context.RequestContext;
 import lombok.extern.slf4j.Slf4j;
 
-import javax.servlet.http.HttpServletRequest;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -51,14 +50,7 @@ public class LocationFilter extends ZuulFilter implements RoutedServicesUser {
 
         final String serviceId = (String) context.get(SERVICE_ID_KEY);
         final String proxy = UrlUtils.removeFirstAndLastSlash((String) context.get(PROXY_KEY));
-        final String requestPath;
-
-        final HttpServletRequest req = context.getRequest();
-        if (req != null) {
-            requestPath = context.getRequest().getRequestURI().replaceFirst("/" + proxy,"");
-        } else {
-            requestPath = UrlUtils.addFirstSlash((String) context.get(REQUEST_URI_KEY));
-        }
+        final String requestPath = UrlUtils.addFirstSlash((String) context.get(REQUEST_URI_KEY));
 
         if (isRequestThatCanBeProcessed(serviceId, proxy, requestPath)) {
             RoutedServices routedServices = routedServicesMap.get(serviceId);
